many minor fixes
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/msgl@43 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/src/include/msgl/msgl_mg_loss.h
@@ -30,7 +30,7 @@ public:
 
 
 	MultinomialSampleLoss() :
-			n_samples(0), n_classes(0), Y(static_cast<u32>(0)), W(0), prob(n_samples, n_classes), hessian_matrices(static_cast<u32>(0)), hessians_computed(false) {
+			n_samples(0), n_classes(0), Y(static_cast<u32>(0)), W(static_cast<u32>(0)), prob(n_samples, n_classes), hessian_matrices(static_cast<u32>(0)), hessians_computed(false) {
 	}
 
 	template<typename E>
